@@ -3,17 +3,17 @@ import 'package:provider/provider.dart';
 import 'package:flutter_store/utils/app_string.dart';
 import 'package:flutter_store/utils/app_routes.dart';
 
-import 'package:flutter_store/providers/cart.dart';
-import 'package:flutter_store/providers/auth.dart';
-import 'package:flutter_store/providers/orders.dart';
-import 'package:flutter_store/providers/products.dart';
-
 import 'package:flutter_store/views/cart_page.dart';
 import 'package:flutter_store/views/orders_page.dart';
-import 'package:flutter_store/views/product_form.dart';
 import 'package:flutter_store/views/products_page.dart';
 import 'package:flutter_store/views/auth_home_page.dart';
+import 'package:flutter_store/views/product_form_page.dart';
 import 'package:flutter_store/views/product_detail_page.dart';
+
+import 'package:flutter_store/providers/auth_provider.dart';
+import 'package:flutter_store/providers/cart_provider.dart';
+import 'package:flutter_store/providers/orders_provider.dart';
+import 'package:flutter_store/providers/products_provider.dart';
 
 void main() => runApp(MyApp());
 
@@ -28,28 +28,31 @@ class MyApp extends StatelessWidget {
         ChangeNotifierProxyProvider<Auth, ProductsProvider>(
           create: (_) => new ProductsProvider(),
           update: (ctx, auth, previousProducts) => ProductsProvider(
-            auth.token,
-            auth.userId,
-            previousProducts.items,
+            auth.token ?? '',
+            auth.userId ?? '',
+            previousProducts?.items ?? [],
           ),
         ),
         ChangeNotifierProvider(
-          create: (ctx) => Cart(),
+          create: (ctx) => CartProvider(),
         ),
-        ChangeNotifierProxyProvider<Auth, Orders>(
-          create: (_) => new Orders(),
-          update: (ctx, auth, previousOrders) => Orders(
-            auth.token,
-            auth.userId,
-            previousOrders.orders,
+        ChangeNotifierProxyProvider<Auth, OrderProvider>(
+          create: (_) => new OrderProvider(),
+          update: (ctx, auth, previousOrders) => OrderProvider(
+            auth.token ?? '',
+            auth.userId ?? '',
+            previousOrders?.items ?? [],
           ),
         ),
       ],
       child: MaterialApp(
         title: AppString.appTitle,
         theme: ThemeData(
-          primarySwatch: Colors.purple,
-          accentColor: Colors.deepOrange,
+          primaryColor: Colors.purple,
+          colorScheme: ColorScheme.fromSwatch(
+            primarySwatch: Colors.purple,
+            accentColor: Colors.deepOrange,
+          ),
           fontFamily: 'Lato',
           elevatedButtonTheme: ElevatedButtonThemeData(
             style: ElevatedButton.styleFrom(
@@ -63,7 +66,6 @@ class MyApp extends StatelessWidget {
             ),
           ),
         ),
-        // home: ProductOverviewPage(),
         debugShowCheckedModeBanner: false,
         routes: {
           AppRoutes.AUTH_HOME: (ctx) => AuthHomePage(),
@@ -71,7 +73,7 @@ class MyApp extends StatelessWidget {
           AppRoutes.CART: (ctx) => CartPage(),
           AppRoutes.ORDERS: (ctx) => OrderPage(),
           AppRoutes.PRODUCTS: (ctx) => ProductsPage(),
-          AppRoutes.PRODUCT_FORM: (ctx) => ProductForm(),
+          AppRoutes.PRODUCT_FORM: (ctx) => ProductFormPage(),
         },
       ),
     );
