@@ -1,58 +1,86 @@
+import 'dart:math';
 import 'dart:convert';
 import 'package:flutter/cupertino.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_store/utils/app_urls.dart';
+import 'package:flutter_store/utils/app_string.dart';
 import 'package:flutter_store/models/product_model.dart';
+import 'package:flutter_store/exceptions/http_exception.dart';
 
 class ProductsProvider with ChangeNotifier {
   final String _baseUrl = '${AppUrl.BASE_API}/products';
   List<ProductModel> _items = [];
-  String _token;
-  String _userId;
+  final String? _token;
+  final String? _userId;
 
   ProductsProvider([this._token, this._userId, this._items = const []]);
 
-  List<ProductModel> get items => [..._items]; // Return copy of Items.
+  List<ProductModel> get items => [..._items]; // Return copy/clone of Items.
+
+  int get itemsCount {
+    return _items.length;
+  }
 
   List<ProductModel> get favoriteItems {
     return _items.where((element) => element.isFavorite).toList();
   } // Return copy of Items.
 
   Future<void> loadProducts() async {
-    final response = await http.get('$_baseUrl.json?auth=$_token');
-    final favoriteResponse = await http
-        .get('${AppUrl.BASE_API}/userFavorites/$_userId.json?auth=$_token');
-    final favoriteMap = jsonDecode(favoriteResponse.body);
-
-    Map<String, dynamic> data = json.decode(response.body);
     _items.clear();
 
-    if (data != null) {
-      data.forEach((key, value) {
-        final isFavorite =
-            favoriteMap == null ? false : favoriteMap[key] ?? false;
-        _items.add(
-          ProductModel(
-            id: key,
-            title: value['title'],
-            price: value['price'],
-            description: value['description'],
-            imageUrl: value['imageUrl'],
-            isFavorite: isFavorite,
-          ),
-        );
-      });
-      notifyListeners();
-    }
+    final response = await http.get(
+      Uri.parse('$_baseUrl.json?auth=$_token'),
+    );
+    final favoriteResponse = await http.get(
+      Uri.parse('${AppUrl.BASE_API}/userFavorites/$_userId.json?auth=$_token'),
+    );
+    final favoriteMap = jsonDecode(favoriteResponse.body);
+
+    if (response.body == 'null') return;
+    Map<String, dynamic> data = json.decode(response.body);
+
+    data.forEach((key, value) {
+      final isFavorite =
+          favoriteMap == null ? false : favoriteMap[key] ?? false;
+      _items.add(
+        ProductModel(
+          id: key,
+          name: value['name'],
+          price: value['price'] as double,
+          description: value['description'],
+          imageUrl: value['imageUrl'],
+          isFavorite: isFavorite,
+        ),
+      );
+    });
+    notifyListeners();
     return Future.value();
+  }
+
+  Future<void> saveProduct(Map<String, Object> data) {
+    bool hasId = data['id'] != null;
+
+    final product = ProductModel(
+      id: hasId ? data['id'] as String : Random().nextDouble().toString(),
+      name: data['name'] as String,
+      price: data['price'] as double,
+      description: data['description'] as String,
+      imageUrl: data['imageUrl'] as String,
+    );
+
+    if (hasId) {
+      return updateProduct(product);
+    } else {
+      return addProduct(product);
+    }
   }
 
   Future<void> addProduct(ProductModel product) async {
     final response = await http.post(
-      '$_baseUrl.json?auth=$_token',
+      Uri.parse('$_baseUrl.json?auth=$_token'),
       body: json.encode(
         {
-          'title': product.title,
+          'name': product.name,
           'description': product.description,
           'price': product.price,
           'imageUrl': product.imageUrl,
@@ -63,7 +91,7 @@ class ProductsProvider with ChangeNotifier {
     _items.add(
       ProductModel(
         id: json.decode(response.body)['name'],
-        title: product.title,
+        name: product.name,
         price: product.price,
         description: product.description,
         imageUrl: product.imageUrl,
@@ -72,18 +100,14 @@ class ProductsProvider with ChangeNotifier {
     notifyListeners(); // Notify components
   }
 
-  Future<void> updateProducts(ProductModel product) async {
-    if (product == null || product.id == null) {
-      return;
-    }
-
+  Future<void> updateProduct(ProductModel product) async {
     final index = _items.indexWhere((item) => item.id == product.id);
 
     if (index >= 0) {
       await http.patch(
-        '$_baseUrl/${product.id}.json?auth=$_token',
+        Uri.parse('$_baseUrl/${product.id}.json?auth=$_token'),
         body: json.encode({
-          'title': product.title,
+          'name': product.name,
           'description': product.description,
           'price': product.price,
           'imageUrl': product.imageUrl,
@@ -96,23 +120,24 @@ class ProductsProvider with ChangeNotifier {
 
   Future<void> deleteProduct(String id) async {
     final index = _items.indexWhere((item) => item.id == id);
+
     if (index >= 0) {
       final product = _items[index];
       _items.remove(product);
       notifyListeners();
 
-      final response =
-          await http.delete('$_baseUrl/${product.id}.json?auth=$_token');
+      final response = await http.delete(
+        Uri.parse('$_baseUrl/${product.id}.json?auth=$_token'),
+      );
       if (response.statusCode >= 400) {
         _items.insert(index, product);
         notifyListeners();
-        // throw HttpException('Erro ao excluir produto.');
+        throw HttpException(
+          msg: AppString.snackBarTextDeleteProdcutError,
+          statusCode: response.statusCode,
+        );
       }
     }
-  }
-
-  int get itemsCount {
-    return _items.length;
   }
 
   // bool _showFavoriteOnly = false;
