@@ -1,10 +1,14 @@
 class HttpException implements Exception {
   final String msg;
+  final int statusCode;
 
-  const HttpException(this.msg);
+  const HttpException({
+    required this.msg,
+    required this.statusCode,
+  });
 
   @override
   String toString() {
-    return this.msg;
+    return msg;
   }
 }
