@@ -8,11 +8,24 @@ class AppString {
   static String authSignin = 'ENTRAR';
   static String authRegister = 'REGISTRAR';
   static String notConfirmedPassword = 'Senhas diferentes.';
+  static String labelOk = 'OK';
+  static String labelName = 'Name';
+  static String labelBuy = 'COMPRAR';
+  static String labelTotal = 'Total';
+  static String labelRemove = 'Remover';
   static String labelUndo = 'DESFAZER';
   static String labelEmail = 'E-mail';
+  static String labelUrlImage = 'URL Imagem';
   static String labelPassword = 'Senha';
+  static String labelPrice = 'Preço';
+  static String labelDescription = 'Descrição';
+  static String labelProductUrlInfo = 'Informe URL';
   static String labelConfirmPassword = 'Confirmar senha';
+  static String titleCart = 'Carrinho';
   static String titleAlertError = 'Erro!';
+  static String titleProducts = 'Produtos';
+  static String titleMyOrders = 'Meus Pedidos';
+  static String titleProductForm = 'Formulário Produto';
   static String closeAlertError = 'Fechar';
   static String drawerLabelStore = 'Loja';
   static String drawerLabelOrders = 'Pedidos';
@@ -23,5 +36,14 @@ class AppString {
   static String alertTextRemoveItemCart = 'Remover item do carrinho?';
   static String alertActionLabelConfirm = 'Sim';
   static String alertActionLabelCancel = 'Não';
-  static String snackBarTextDeleteProdcutError = 'Erro ao excluir produto.';
+  static String unexpectedError = 'Erro inesperado';
+  static String requestError = 'Erro de requisição';
+  static String authenticationError = 'Ocorreu um erro durante atenticação.';
+  static String caractereValidator = 'Mínimo de 3 caracteres.';
+  static String invalidNameValidator = 'Informe nome.';
+  static String invalidPriceValidator = 'Informe um preço válido.';
+  static String invalidImageValidator = 'Informar uma url válida.';
+  static String invalidDescriptionValidator = 'Informar descrição';
+  static String snackBarTextDeleteProdcutError =
+      'Não foi possível excluir o produto.';
 }
