@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_store/utils/app_routes.dart';
 import 'package:provider/provider.dart';
+import 'package:flutter_store/utils/app_routes.dart';
+import 'package:flutter_store/utils/app_string.dart';
 import 'package:flutter_store/widgets/products_crud.dart';
-import 'package:flutter_store/providers/products.dart';
 import 'package:flutter_store/widgets/app_drawer_widget.dart';
+import 'package:flutter_store/providers/products_provider.dart';
 
 class ProductsPage extends StatelessWidget {
   Future<void> _refreshProducts(BuildContext context) async {
@@ -14,10 +15,11 @@ class ProductsPage extends StatelessWidget {
   Widget build(BuildContext context) {
     final productsProvider = Provider.of<ProductsProvider>(context);
     final products = productsProvider.items;
+
     return Scaffold(
       drawer: AppDrawerWidget(),
       appBar: AppBar(
-        title: Text('Produtos'),
+        title: Text(AppString.titleProducts),
         actions: <Widget>[
           IconButton(
             icon: const Icon(Icons.add),
@@ -30,12 +32,12 @@ class ProductsPage extends StatelessWidget {
       body: RefreshIndicator(
         onRefresh: () => _refreshProducts(context),
         child: Padding(
-          padding: EdgeInsets.all(8.0),
+          padding: const EdgeInsets.all(8.0),
           child: ListView.builder(
             itemCount: productsProvider.itemsCount,
             itemBuilder: (ctx, index) => Column(
               children: <Widget>[
-                ProductCrud(products[index]),
+                ProductCrudWidget(products[index]),
                 Divider(),
               ],
             ),
