@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_store/exceptions/auth_exception.dart';
 import 'package:provider/provider.dart';
-import 'package:flutter_store/providers/auth.dart';
 import 'package:flutter_store/utils/app_string.dart';
+import 'package:flutter_store/providers/auth_provider.dart';
+import 'package:flutter_store/exceptions/auth_exception.dart';
 
 enum AuthMode { Signup, Login }
 
@@ -17,39 +17,44 @@ class _AuthCardWidgetState extends State<AuthCardWidget> {
     'password': '',
   };
 
-  GlobalKey<FormState> _form = GlobalKey();
   bool _isLoading = false;
   AuthMode _authMode = AuthMode.Login;
+  GlobalKey<FormState> _form = GlobalKey();
   final _passwordController = TextEditingController();
 
+  bool _isLogin() => _authMode == AuthMode.Login;
+  bool _isSignup() => _authMode == AuthMode.Signup;
+
   Future<void> _submit() async {
-    if (!_form.currentState.validate()) {
+    if (!_form.currentState!.validate()) {
       return;
     }
 
-    setState(() {
-      _isLoading = true;
-    });
+    setState(() => _isLoading = true);
 
-    _form.currentState.save();
+    _form.currentState?.save();
 
     Auth auth = Provider.of(context, listen: false);
 
     try {
-      if (_authMode == AuthMode.Login) {
-        await auth.login(_authData["email"], _authData["password"]);
+      if (_isLogin()) {
+        await auth.login(
+          _authData["email"] as String,
+          _authData["password"] as String,
+        );
       } else {
-        await auth.signup(_authData["email"], _authData["password"]);
+        await auth.signup(
+          _authData["email"] as String,
+          _authData["password"] as String,
+        );
       }
     } on AuthException catch (error) {
       _showErrorDialog(error.toString());
     } catch (error) {
-      _showErrorDialog('Erro inesperado.');
+      _showErrorDialog(AppString.unexpectedError);
     }
 
-    setState(() {
-      _isLoading = false;
-    });
+    setState(() => _isLoading = false);
   }
 
   void _showErrorDialog(String message) {
@@ -71,14 +76,10 @@ class _AuthCardWidgetState extends State<AuthCardWidget> {
   }
 
   void _switchAuthMode() {
-    if (_authMode == AuthMode.Login) {
-      setState(() {
-        _authMode = AuthMode.Signup;
-      });
+    if (_isLogin()) {
+      setState(() => _authMode = AuthMode.Signup);
     } else {
-      setState(() {
-        _authMode = AuthMode.Login;
-      });
+      setState(() => _authMode = AuthMode.Login);
     }
   }
 
@@ -89,10 +90,10 @@ class _AuthCardWidgetState extends State<AuthCardWidget> {
     return Card(
       elevation: 8.0,
       shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(10),
+        borderRadius: BorderRadius.circular(10.0),
       ),
       child: Container(
-        height: _authMode == AuthMode.Login ? 300 : 370,
+        height: _isLogin() ? 300.0 : 370.0,
         width: deviceSize.width * 0.75,
         padding: EdgeInsets.all(16.0),
         child: Form(
@@ -105,12 +106,12 @@ class _AuthCardWidgetState extends State<AuthCardWidget> {
                 ),
                 keyboardType: TextInputType.emailAddress,
                 validator: (value) {
-                  if (value.isEmpty || !value.contains('@')) {
+                  if (value!.isEmpty || !value.contains('@')) {
                     return AppString.authInvalidEmail;
                   }
                   return null;
                 },
-                onSaved: (value) => _authData['email'] = value,
+                onSaved: (email) => _authData['email'] = email ?? '',
               ),
               TextFormField(
                 decoration: InputDecoration(
@@ -120,28 +121,28 @@ class _AuthCardWidgetState extends State<AuthCardWidget> {
                 obscureText: true,
                 controller: _passwordController,
                 validator: (value) {
-                  if (value.isEmpty || value.length < 5) {
+                  if (value!.isEmpty || value.length < 5) {
                     return AppString.authInvalidPassword;
                   }
                   return null;
                 },
-                onSaved: (value) => _authData['password'] = value,
+                onSaved: (password) => _authData['password'] = password ?? '',
               ),
-              if (_authMode == AuthMode.Signup)
+              if (_isSignup())
                 TextFormField(
                   decoration: InputDecoration(
                     labelText: AppString.labelConfirmPassword,
                   ),
                   keyboardType: TextInputType.text,
                   obscureText: true,
-                  validator: _authMode == AuthMode.Signup
-                      ? (value) {
+                  validator: _isLogin()
+                      ? null
+                      : (value) {
                           if (value != _passwordController.text) {
                             return AppString.notConfirmedPassword;
                           }
                           return null;
-                        }
-                      : null,
+                        },
                 ),
               Spacer(),
               _isLoading
@@ -153,6 +154,15 @@ class _AuthCardWidgetState extends State<AuthCardWidget> {
                             : AppString.authRegister,
                       ),
                       onPressed: _submit,
+                      style: ElevatedButton.styleFrom(
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(30.0),
+                        ),
+                        padding: const EdgeInsets.symmetric(
+                          horizontal: 30.0,
+                          vertical: 8.0,
+                        ),
+                      ),
                     ),
               TextButton(
                 onPressed: _switchAuthMode,
