@@ -2,24 +2,26 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_store/utils/app_string.dart';
 import 'package:flutter_store/utils/app_routes.dart';
-import 'package:flutter_store/providers/products.dart';
 import 'package:flutter_store/models/product_model.dart';
-// import 'package:flutter_store/exceptions/http_exception.dart';
+import 'package:flutter_store/exceptions/http_exception.dart';
+import 'package:flutter_store/providers/products_provider.dart';
 
-class ProductCrud extends StatelessWidget {
+class ProductCrudWidget extends StatelessWidget {
   final ProductModel productModel;
 
-  ProductCrud(this.productModel);
+  ProductCrudWidget(this.productModel);
 
   @override
   Widget build(BuildContext context) {
     final scaffold = ScaffoldMessenger.of(context);
+
     return ListTile(
-      leading:
-          CircleAvatar(backgroundImage: NetworkImage(productModel.imageUrl)),
-      title: Text(productModel.title),
+      leading: CircleAvatar(
+        backgroundImage: NetworkImage(productModel.imageUrl),
+      ),
+      title: Text(productModel.name),
       trailing: Container(
-        width: 100,
+        width: 100.0,
         child: Row(
           children: <Widget>[
             IconButton(
@@ -40,7 +42,8 @@ class ProductCrud extends StatelessWidget {
                   context: context,
                   builder: (ctx) => AlertDialog(
                     title: Text(AppString.alertTitleConfirmDelete),
-                    content: Text("Remover \'${productModel.title}'?"),
+                    content: Text(
+                        "${AppString.labelRemove} \'${productModel.name}'?"),
                     actions: <Widget>[
                       TextButton(
                         child: Text(AppString.alertActionLabelConfirm),
@@ -59,14 +62,12 @@ class ProductCrud extends StatelessWidget {
                         await Provider.of<ProductsProvider>(
                           context,
                           listen: false,
-                        ).deleteProduct(
-                          productModel.id,
-                        );
-                      } catch (error) {
+                        ).deleteProduct(productModel.id);
+                      } on HttpException catch (error) {
                         scaffold.showSnackBar(
                           SnackBar(
                             content: Text(
-                              AppString.snackBarTextDeleteProdcutError,
+                              error.toString(),
                             ),
                           ),
                         );
